// resources/scripts/interfaces/VacationPeriodInterface.ts
export interface VacationPeriodInterface {
  vacationPeriodRange: string
  vacationPeriodYear: number
  vacationSettingId: number
  vacationPeriodActiveWorkYears: number
  vacationPeriodCorrespondingDays: number
  vacationPeriodUsedDays: number
  vacationPreviousAvailableDays: number
  vacationPeriodAvailableDays: number
}